Start working on ConfiguredSurfaceBuilder
