+++ src/main/java/com/swordglowsblue/artifice/api/builder/data/dimension/GeneratorSettingsBuilder.java
@@ -2,6 +2,7 @@ package com.swordglowsblue.artifice.api.builder.data.dimension;
 
 import com.google.gson.JsonObject;
 import com.swordglowsblue.artifice.api.builder.TypedJsonBuilder;
+import com.swordglowsblue.artifice.api.builder.data.BlockStateBuilder;
 import com.swordglowsblue.artifice.api.util.Processor;
 
 public class GeneratorSettingsBuilder extends TypedJsonBuilder<JsonObject> {
@@ -95,37 +96,5 @@ public class GeneratorSettingsBuilder extends TypedJsonBuilder<JsonObject> {
     public GeneratorSettingsBuilder structureManager(Processor<StructureManagerBuilder> structureManagerBuilder) {
         with("structures", JsonObject::new, jsonObject -> structureManagerBuilder.process(new StructureManagerBuilder()).buildTo(jsonObject));
         return this;
-    }
-
-
-    public static class BlockStateBuilder extends TypedJsonBuilder<JsonObject> {
-
-        private JsonObject jsonObject = new JsonObject();
-
-        protected BlockStateBuilder() {
-            super(new JsonObject(), j->j);
-        }
-
-        /**
-         * Set the id of the block.
-         * @param id
-         * @return
-         */
-        public BlockStateBuilder name(String id) {
-            this.root.addProperty("Name", id);
-            return this;
-        }
-
-        /**
-         * Set a property to a state.
-         * @param property
-         * @param state
-         * @return
-         */
-        public BlockStateBuilder setProperty(String property, String state) {
-            this.jsonObject.addProperty(property, state);
-            this.root.add("Properties", this.jsonObject);
-            return this;
-        }
     }
 }
